+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Unreleased (2.0.0)
 
 - Removed deprecated `AnalysisOptions.from` use `AnalysisOptions.fromMap` instead
+- Removed deprecated `Config.linesOfCodeWarningLevel` use `Config.linesOfExecutableCodeWarningLevel` instead
 
 # 1.10.0